weather/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
requests==2.32.3
celery==5.4.0
gunicorn==20.1.0
redis==5.0.4
django-celery-beat==2.6.0
django-celery-results==2.5.1
psycopg2-binary==2.9.1
dj-rest-auth==6.0.0
python-dotenv==1.0.1
